Moved IgnoredTypeAdapter to adapters directory
