+++ hive/lib/src/adapters/ignored_type_adapter.dart
