+++ hive/test/tests/adapters/ignored_type_adapter_test.dart
@@ -1,4 +1,4 @@
-import 'package:hive/src/registry/ignored_type_adapter.dart';
+import 'package:hive/src/adapters/ignored_type_adapter.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
